// src/common/enums.ts
export enum LbMode {
    Alltime = '0',
    Week = '1'
}

export enum FriendRequestStatus {
    Pending = 'PENDING',
    Accepted = 'ACCEPTED',
    Rejected = 'REJECTED'
}

export enum ChatBanLevel {
    NotBanned = 0,
    BannedNotify = 1,
    ShadowBan = 2
}

export enum GameBanLevel {
    NotBanned = 0,
    AntiWin = 1,
    JoinBan = 2
}

export enum ProductionStatus {
    NotStarted = 'notStarted',
    InProgress = 'inProgress',
    Complete = 'complete'
}